+++ lib/SerialComManager/SerialComManager.cpp
@@ -23,7 +23,7 @@ SerialComManager::SerialComManager(
 void SerialComManager::receiveSerialData()
 {
     static String serialPortData = "";
-    uint8_t c = "";
+    uint16_t c = "";
 
     // If we didn't receive some data from ESP for a certain amount of time, we stop the car to be safe
     if (timeManager->getLoopTime() - lastReceiveTime > ESP_DATA_MAX_RECEIVE_INTERVAL)
